consolidated df-valve sub projects into one / demolition happening in df-auth
