+++ df-auth/factory/src/main/java/com/hashmapinc/tempus/witsml/auth/AbstractAuth.java
@@ -1,0 +1,30 @@
+/**
+ * Copyright © 2018-2018 Hashmap, Inc
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.hashmapinc.tempus.witsml.auth;
+
+public abstract class AbstractAuth {
+    /**
+     * Retrieve the name of the valve
+     * @return The name of the valve
+     */
+    public abstract String getName();
+
+    /**
+     * Retrieve the name of the valve
+     * @return The name of the valve
+     */
+    public abstract String getDescription();
+}
