+++ df-auth/factory/src/main/java/com/hashmapinc/tempus/witsml/auth/AuthFactory.java
@@ -13,8 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hashmapinc.tempus.witsml.valve;
+package com.hashmapinc.tempus.witsml.auth;
 
-public class ObjectValve {
+import com.hashmapinc.tempus.witsml.auth.AbstractAuth;
 
+public class AuthFactory {
+    /**
+     * creates an auth object based on the authType requested
+     * 
+     * @param authType - String name of the auth implementation to use
+     * 
+     */
+    public static AbstractAuth buildValve(String valveType) {
+        switch (valveType) {
+            case "DoT":
+                return new com.hashmapinc.tempus.witsml.auth.dot.DotAuth();
+            default:
+                return null;
+        }
+    }
 }
