+++ df-valve/pom.xml
@@ -43,7 +43,7 @@
     </parent>
 
     <artifactId>df-valve</artifactId>
-    <packaging>pom</packaging>
+    <packaging>jar</packaging>
 
     <name>DrillFlow Valve Module</name>
     <url>https://hashmapinc.com</url>
@@ -53,8 +53,12 @@
         <main.dir>${basedir}/..</main.dir>
     </properties>
 
-    <modules>
-        <module>dot-valve</module>
-        <module>factory</module>
-    </modules>
+    <dependencies>
+        <!-- used for witsml object ser/de -->
+        <dependency>
+            <groupId>com.hashmapinc.tempus</groupId>
+            <artifactId>WitsmlObjects</artifactId>
+            <version>1.1.1</version>
+        </dependency>
+    </dependencies>
 </project>
