+++ df-valve/src/main/java/com/hashmapinc/tempus/witsml/valve/dot/DotTranslator.java
