+++ df-valve/src/main/java/com/hashmapinc/tempus/witsml/valve/mock/MockDelegator.java
